Break map schema out from TagFormat.  Further refine hed2->hed3 upgrade functions
