+++ hedtools/tests/tests_tag_format.py
@@ -11,7 +11,7 @@ class Test(unittest.TestCase):
     def setUpClass(cls):
         hed_xml = os.path.join(os.path.dirname(os.path.abspath(__file__)), cls.schema_file)
         cls.tag_compare = TagFormat(hed_xml)
-        cls.tag_compare.print_tag_dict()
+        cls.tag_compare.map_schema.print_tag_dict()
 
     def compare_base_new(self, test_function, input_strings, expected_results, errors_list=None):
         # Assume there are no errors if none provided.
